# packages/gateway-service/src/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Authenticate user and return access token",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login successful",
                    "data": {
                      "accessToken": "string",
                      "refreshToken": "string",
                      "user": {
                        "id": "number",
                        "username": "string",
                        "email": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing credentials"
          },
          "401": {
            "description": "Unauthorized - invalid credentials"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Register a new user account",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Registration successful",
                    "data": {
                      "user": {
                        "id": "number",
                        "username": "string",
                        "email": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields or invalid data"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User registration",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "description": "Get new access token using refresh token",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Token refreshed successfully",
                    "data": {
                      "accessToken": "string",
                      "refreshToken": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - refresh token missing"
          },
          "401": {
            "description": "Unauthorized - invalid refresh token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    }
  },
  "info": {
    "title": "API文档",
    "description": "系统API接口文档",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username",
            "description": "The username of the user"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": ""
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username",
            "description": "The username of the user"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": "The password of the user"
          },
          "email": {
            "type": "string",
            "example": "email",
            "description": "The email of the user"
          }
        },
        "required": [
          "username",
          "password",
          "email"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "refreshToken",
            "description": "The refresh token of the user"
          }
        },
        "required": [
          "refreshToken"
        ]
      }
    }
  }
}